Update Home Page HTML and CSS
This commit changes the display of the table and allows for the table traversal through pages. Currently each page displays ten devices. This could be updated going forward, depending on the needs. A step forward is to separate the HTML, CSS and JavaScript used for the table, such that it can easily be imported and reused for other pages.