Update Device Status
The status of a device has been changed from a boolean value to an Enum which has values ONLINE, WARNING, OFFLINE. The code has been changed such that it works with this datatype. The Device List page and its table have also been updated such that the data representation and functionality remain the same.