Fix Timestamp Issues
There were issues with comparing timestamps since the devices were sending UTC time, while it was comparing it to Local Time. Another issue was with the statuses being displayed, this has also been resolved. One thing to note is that the overview page does not automatically refresh every 60 seconds, nor does it query the server again, so the device statuses are not updated.