Update DeviceList
The devise list page now uses pagination for the table. However, the search and filtering are not fully functional yet when integrated with pagination. For example, if we search for devices, and then want to view the second page of the search result, it resets the table. Searching also does decrease the number of pages.